+++ CardRouletteTests/CardRouletteTests.swift
@@ -8,29 +8,17 @@
 
 import UIKit
 import XCTest
+@testable import CardRoulette
 
 class CardRouletteTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
+
+    func testParticipantNameNormalizationTrimsWhitespace() {
+        XCTAssertEqual(ParticipantListItem.normalizedName("  Hemal\n")!, "Hemal", "Participant names should be trimmed before saving")
     }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
+
+    func testParticipantNameNormalizationRejectsBlankNames() {
+        XCTAssertNil(ParticipantListItem.normalizedName("  \n\t  "), "Blank participant names should be ignored")
+        XCTAssertNil(ParticipantListItem.normalizedName(nil), "Missing participant names should be ignored")
     }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        XCTAssert(true, "Pass")
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock() {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
+
 }